requirements.txt
Django==1.6.6
Pillow==2.4.0
argparse==1.2.1
bcrypt==1.0.2
certifi==14.05.14
cffi==0.8.6
django-jsonfield==0.9.12b2
django-oauth2-provider==0.2.6.1
django-password-reset==0.7
psycopg2==2.5.1
py-bcrypt==0.4
pycparser==2.10
pytz==2014.3
requests==2.4.0
shortuuid==0.4.2
six==1.7.3
wsgiref==0.1.2
